Initial commit.
This is a long story.

Two years ago, when I tried to build a new personal website, I wanted to
find an AWESOME static site generator. After googling serveral days,
finally I met nanoc and fall in love with it.

Nanoc is really a wonderful and flexible tool to build static websites.
However, it has a deep learning curve compared with other static site
generators. In my opinion, nanoc is a generic web content generator and
manager, you can use it to build all parts of a small-to-medium site if
you don't need backend server.

I've used nanoc to build multiple static web sites and I still love its
flexibility. So I think it's time for me to release a new bundle built
on top of nanoc to make it more smooth to build a website using nanoc.

That's where this project starts:

- nanoc 4.2.4
- pandoc
- pandoc-ruby
- semantic-ui 2.2.2
- fonts.css 1.4.0
- jquery 3.0.0
- linux libertine 5.2.8