Components: dashboard
The navbar can be fixed on the top, but the left vertical menu can not
at present. Sidebar and sticky functions are not working very well.

Sidebar: it will push column on the right out of the viewport, instead
of compress it, so I do not add JavaScript for now.

Sticky: not very clear about how it works, because I put sidebar
and main content in one row, it does not work as expected.