Experiments: off-canvas
Off canvas basically means something outside the viewport. Because
Semantic-UI's sidebar has best performance as a direct children of the
 body, here it's not so perfect displayed, especially when it's set to
 be visible on loading.

So I have decided to make two versions of page content, one for computer,
and one for smaller screens, by which it seems similar implemantation of
the Bootstrap sidebar, but actually it's not. Semantic-UI's sidebar has
advantages in other forms.