Add "Preview/Download" links to templates.
We provide a "Preview" button link to each template, when user click on
"Preview" button, it will jump to the live template preview page.

And in addition to the explicitly "Preview" button, when user click on
the image screenshot of each template, it will also jump to the live
template preview page.

Besides "Preview" button, we provide a "Download" button for users,
where users can get a offline tarball for offline web page viewing.

We use `wget` to convert page links to offline links.

Ref:
- https://www.labnol.org/software/wget-command-examples/28750/